+++ packages/terra-form-checkbox/src/CheckboxField.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import uniqueid from 'lodash.uniqueid';
 import VisualyHiddenText from 'terra-visually-hidden-text';
-import { injectIntl, intlShape } from 'react-intl';
+
 import styles from './CheckboxField.module.scss';
 
 const cx = classNames.bind(styles);
@@ -25,11 +25,6 @@ const propTypes = {
    * Whether or not to hide the required indicator on the legend.
    */
   hideRequired: PropTypes.bool,
-  /**
-   * @private
-   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
-   */
-  intl: intlShape.isRequired,
   /**
    * Whether or not the field is an inline field.
    */
@@ -74,13 +69,21 @@ const defaultProps = {
   showOptional: false,
 };
 
-const CheckboxField = (props) => {
+const contextTypes = {
+  /* eslint-disable consistent-return */
+  intl: (context) => {
+    if (context.intl === undefined) {
+      return new Error('Component is internationalized, and must be wrapped in terra-base');
+    }
+  },
+};
+
+const CheckboxField = (props, { intl }) => {
   const {
     children,
     error,
     help,
     hideRequired,
-    intl,
     isInvalid,
     isInline,
     isLegendHidden,
@@ -147,5 +150,6 @@ const CheckboxField = (props) => {
 
 CheckboxField.propTypes = propTypes;
 CheckboxField.defaultProps = defaultProps;
+CheckboxField.contextTypes = contextTypes;
 
-export default injectIntl(CheckboxField);
+export default CheckboxField;