+++ packages/terra-form-select/tests/jest/Select.test.jsx
@@ -7,102 +7,102 @@ import OptGroup from '../../src/_OptGroup';
 
 describe('Select', () => {
   it('should render a default variant', () => {
-    const wrapper = shallowWithIntl(<Select />).dive();
+    const wrapper = shallowWithIntl(<Select />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a combobox variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="tag" />).dive();
+    const wrapper = shallowWithIntl(<Select variant="combobox" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a multiple variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="multiple" />).dive();
+    const wrapper = shallowWithIntl(<Select variant="multiple" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a search variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="search" />).dive();
+    const wrapper = shallowWithIntl(<Select variant="search" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a tag variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="tag" />).dive();
+    const wrapper = shallowWithIntl(<Select variant="tag" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a default variant with a placeholder', () => {
-    const wrapper = shallowWithIntl(<Select placeholder="Placeholder" />).dive();
+    const wrapper = shallowWithIntl(<Select placeholder="Placeholder" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a combobox variant with a placeholder', () => {
-    const wrapper = shallowWithIntl(<Select variant="tag" placeholder="Placeholder" />).dive();
+    const wrapper = shallowWithIntl(<Select variant="combobox" placeholder="Placeholder" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a multiple variant with a placeholder', () => {
-    const wrapper = shallowWithIntl(<Select variant="multiple" placeholder="Placeholder" />).dive();
+    const wrapper = shallowWithIntl(<Select variant="multiple" placeholder="Placeholder" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a search variant with a placeholder', () => {
-    const wrapper = shallowWithIntl(<Select variant="search" placeholder="Placeholder" />).dive();
+    const wrapper = shallowWithIntl(<Select variant="search" placeholder="Placeholder" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a tag variant with a placeholder', () => {
-    const wrapper = shallowWithIntl(<Select variant="tag" placeholder="Placeholder" />).dive();
+    const wrapper = shallowWithIntl(<Select variant="tag" placeholder="Placeholder" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a disabled default variant', () => {
-    const wrapper = shallowWithIntl(<Select disabled />).dive();
+    const wrapper = shallowWithIntl(<Select disabled />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a disabled combobox variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="tag" disabled />).dive();
+    const wrapper = shallowWithIntl(<Select variant="tag" disabled />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a disabled multiple variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="multiple" disabled />).dive();
+    const wrapper = shallowWithIntl(<Select variant="multiple" disabled />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a disabled search variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="search" disabled />).dive();
+    const wrapper = shallowWithIntl(<Select variant="search" disabled />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a disabled tag variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="tag" disabled />).dive();
+    const wrapper = shallowWithIntl(<Select variant="tag" disabled />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an invalid default variant', () => {
-    const wrapper = shallowWithIntl(<Select isInvalid />).dive();
+    const wrapper = shallowWithIntl(<Select isInvalid />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an invalid combobox variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="tag" isInvalid />).dive();
+    const wrapper = shallowWithIntl(<Select variant="combobox" isInvalid />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an invalid multiple variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="multiple" isInvalid />).dive();
+    const wrapper = shallowWithIntl(<Select variant="multiple" isInvalid />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an invalid search variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="search" isInvalid />).dive();
+    const wrapper = shallowWithIntl(<Select variant="search" isInvalid />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render an invalid tag variant', () => {
-    const wrapper = shallowWithIntl(<Select variant="tag" isInvalid />).dive();
+    const wrapper = shallowWithIntl(<Select variant="tag" isInvalid />);
     expect(wrapper).toMatchSnapshot();
   });
 
@@ -152,7 +152,7 @@ describe('Select', () => {
 
   it('should call onBlur', () => {
     const mockBlur = jest.fn();
-    const wrapper = shallowWithIntl(<Select onBlur={mockBlur} />).dive();
+    const wrapper = shallowWithIntl(<Select onBlur={mockBlur} />);
 
     wrapper.simulate('focus');
     wrapper.simulate('blur');
@@ -162,7 +162,7 @@ describe('Select', () => {
 
   it('should call onFocus', () => {
     const mockFocus = jest.fn();
-    const wrapper = shallowWithIntl(<Select onFocus={mockFocus} />).dive();
+    const wrapper = shallowWithIntl(<Select onFocus={mockFocus} />);
 
     wrapper.simulate('focus');
 
@@ -171,7 +171,7 @@ describe('Select', () => {
 
   it('should call onClick', () => {
     const mockClick = jest.fn();
-    const wrapper = shallowWithIntl(<Select onClick={mockClick} />).dive();
+    const wrapper = shallowWithIntl(<Select onClick={mockClick} />);
 
     wrapper.simulate('click');
 
@@ -179,12 +179,12 @@ describe('Select', () => {
   });
 
   it('should render a clear option', () => {
-    const wrapper = shallowWithIntl(<Select allowClear />).dive();
+    const wrapper = shallowWithIntl(<Select allowClear />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a required select', () => {
-    const wrapper = shallowWithIntl(<Select required />).dive();
+    const wrapper = shallowWithIntl(<Select required />);
     expect(wrapper).toMatchSnapshot();
   });
 });