+++ packages/terra-action-header/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+2.23.0 - (July 24, 2019)
+------------------
+### Added
+* Added documentation that maximize/minimize buttons will not be shown with the back button
+
+2.22.0 - (July 17, 2019)
+------------------
+### Changed
+* Revert components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
+
 2.21.0 - (July 17, 2019)
 ------------------
 ### Changed