+++ packages/terra-alert/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+4.6.0 - (July 24, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+4.5.0 - (July 17, 2019)
+------------------
+### Changed
+* Revert components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
+
 4.4.0 - (July 17, 2019)
 ------------------
 ### Changed