+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+3.20.0 - (July 24, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.19.0 - (July 17, 2019)
+------------------
+### Changed
+* Revert components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
+
 3.18.0 - (July 17, 2019)
 ------------------
 ### Changed
@@ -17,7 +27,6 @@ Unreleased
 3.16.0  - (July 11, 2019)
 ------------------
 ### Changed
-* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports
 * Updated Docs around the requirement of terra-base and aggregate-translations.