+++ packages/terra-dialog/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+2.20.0 - (July 24, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.19.0 - (July 17, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.18.0 - (July 17, 2019)
 ------------------
 ### Changed
@@ -17,7 +27,6 @@ Unreleased
 2.16.0  - (July 11, 2019)
 ------------------
 ### Changed
-* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports
 * Updated Docs around the requirement of terra-base and aggregate-translations.