+++ packages/terra-form-field/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+3.18.0 - (July 24, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.17.0 - (July 17, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 3.16.0 - (July 17, 2019)
 ------------------
 ### Changed
@@ -17,7 +27,6 @@ Unreleased
 3.14.0  - (July 11, 2019)
 ------------------
 ### Changed
-* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 * Updated Docs around the requirement of terra-base and aggregate-translations.