+++ packages/terra-form-fieldset/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+2.21.0 - (July 24, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.20.0 - (July 17, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.19.0 - (July 17, 2019)
 ------------------
 ### Changed