+++ packages/terra-form-select/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+5.24.0 - (July 24, 2019)
+------------------
+### Changed
+* Update Spanish translations
+
+5.23.0 - (July 17, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 5.22.0 - (July 17, 2019)
 ------------------
 ### Changed
@@ -14,10 +24,12 @@ Unreleased
 ### Added
 * Added documentation on required peerDependencies
 
+### Fixed
+* Issue with invalid options (options with only spaces) being added to the dropdown menu in combobox variant.
+
 5.20.0  - (July 11, 2019)
 ------------------
 ### Changed
-* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 * Updated Docs around the requirement of terra-base and aggregate-translations.