+++ packages/terra-paginator/CHANGELOG.md
@@ -4,6 +4,29 @@ ChangeLog
 Unreleased
 ----------
 
+2.21.1 - (July 24, 2019)
+------------------
+### Changed
+* Update changelog
+
+2.21.0 - (July 24, 2019)
+------------------
+### Added
+* Added new theme variables - see UPGRADEGUIDE.md
+
+### Fixed
+* Updated anchors to buttons for accessibility
+* Added support for i18n and translations
+
+### Changed
+* Removed theme variables - see UPGRADEGUIDE.md
+* Updated to use terra-visually-hidden-text component
+
+2.20.0 - (July 17, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.19.0 - (July 17, 2019)
 ------------------
 ### Changed
@@ -13,9 +36,16 @@ Unreleased
 ------------------
 ### Added
 * Added documentation on required peerDependencies
+* Added new theme variables - see UPGRADEGUIDE.md
+
+### Fixed
+* Updated anchors to buttons for accessibility
+* Added support for i18n and translations
 
 ### Changed
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+* Removed theme variables - see UPGRADEGUIDE.md
+* Updated to use terra-visually-hidden-text component
 
 2.17.0  - (July 11, 2019)
 ------------------