+++ packages/terra-signature/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.16.0 - (July 24, 2019)
+------------------
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
 2.15.0 - (July 16, 2019)
 ------------------
 ### Added