+++ packages/terra-table/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+3.20.0 - (July 24, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 3.19.0 - (July 16, 2019)
 ------------------
 ### Added
@@ -12,6 +17,7 @@ Unreleased
 3.18.0  - (July 11, 2019)
 ------------------
 ### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 * Increased subheader css specificity
 
 3.17.0 - (June 28, 2019)